+++ client/src/app/shared/inputs/input-radio-a/input-radio-a.component.ts
@@ -98,10 +98,12 @@ export class InputRadioAComponent implements OnInit, OnDestroy {
   }
 
   ngOnDestroy() {
-    let steps = this.stepModelService.getStepsModel();
-    steps = steps.filter(step => step !== this.step);
-    this.stepModelService.init(steps, this.documentType);
-    this.stepModelService.removeStep();
+    if (this.commonsService.isBrowser()) {
+      let steps = this.stepModelService.getStepsModel();
+      steps = steps.filter(step => step !== this.step);
+      this.stepModelService.init(steps, this.documentType);
+      this.stepModelService.removeStep();
+    }
   }
 
   createStep() {
