+++ client/src/app/shared/inputs/input-text/input-text.component.ts
@@ -91,10 +91,12 @@ export class InputTextComponent implements OnInit, OnDestroy {
   }
 
   ngOnDestroy() {
-    let steps = this.stepModelService.getStepsModel();
-    steps = steps.filter(step => step !== this.step);
-    this.stepModelService.init(steps, this.documentType);
-    this.stepModelService.removeStep();
+    if (this.commonsService.isBrowser()) {
+      let steps = this.stepModelService.getStepsModel();
+      steps = steps.filter(step => step !== this.step);
+      this.stepModelService.init(steps, this.documentType);
+      this.stepModelService.removeStep();
+    }
   }
 
   /**NEW FORM**/
